pyproject.toml
[project]
name = "doctor-who-ir"
version = "1.0.0"
description = "Information Retrieval System for Doctor Who TV Series"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["information-retrieval", "nlp", "search", "doctor-who", "bm25", "semantic-search"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "nltk>=3.8",
    "pandas>=1.5.0",
    "rank-bm25>=0.2.0",
    "sentence-transformers>=2.2.0",
    "torch>=1.10.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[tool.uv]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
